refactor: use json managed reference
Previously the output for JSON response was repeated semi-infinitely, this can be tackled by using @JsonBackReference on Tasks where the foreign key points to user.

On the other hand on the Tasks, it has relations between many Tasks to one User so it requires @JsonManagedReference to indicate that Users' tasks are managed by User.